ENH: move 'journal_events' from 'funcs' to 'importer'; add logging to 'importer'.
